SSN Encryption/Decryption (#2624)
* - ADding t2 schema

* - small fixes to naming
- added t2.py
- Updating var names to match field names

* - Added new doc for T2
- Updated model fo rT2
- Added active parse function

* - Added T3 schema defs

* - Fixing most lint errors

* - Updated T3 to multirow record

* - Added unit test and data files
- updated field names

* - Updating var names to conform to DB max var name length

* - Updating based on unit test failure

* - adding datafile

* - Adding unit tests for t2 and t3

* - Fixed field types
- Updated test file

* - Removing many migrations to consolodate into one migration

* Revert "- Removing many migrations to consolodate into one migration"

This reverts commit 1455ae4d334f72e250405803d61d26c0a894e886.

* - Resolved test issues after merge

* - Resolve lint errors

* - Merged 1113.2

* - Removed unnecessary file

* - Rename model fields
- Combined migrations

* - fixed spelling in test

* - Fixed lint errors

* rm commented validators

* - Updated schema newlines to be consistent

* - Updated field names
- Updated tests
- Updated migration

* - consolodating migrations

* - Updated readme and backup script

* - Updated parse logic to batch create records

* - Fixed lint errors

* - Batching record serialization

* - Updated parse logic

* - Adding big s1 datafile

* - fixing lint errors

* - Removing test to see if its causing weird failure

* - Updating based on comments
- Removing big file since circi cant handle it

* - fixing error in update method

* - fixing error in update

* - Removed extraneous seek

* - updating ignore to ignore patch files

* - Made evaluate_trailer pure/immutable

* Revert "- Removing test to see if its causing weird failure"

This reverts commit 64b78b737c97fb9bfa70217ff70ccffea8d85429.

* - Use custom transactions while parsing
- Handle transaction rollback on parse failure/error
- Update tests to use transaction logic
- Created new decorator to rollback database when assertion error is hit
- Making elastic search log level configurable
- Added test for inter parse rollback

* - updated store logic to one liner
- updated decorator to catch all exceptions
- updated tests

* - removed function
- renamed test

* - refactored multirecord schema to be a schema manager instead
- updated parse functions to use the new layout
- updated bulk create to manage batch size

* - Name update for clarity

* - Fix lint errors

* - Changing batch size to avoid crash in circi

* - Updated based on review
- Updated tests to reflect line number

* - Updated based on review/OH
- Added extra transaction safety

* - Updated view to show latest datafiles
- Added admin filter to show newest or all datafile records
- Updated indices to allow easier elastic queries

* - Updated based on review  comments

* - Updated based on feedback for preparse error handling
- updated tests

* - Updated search indices to have parent FK

* - Fix lint errors

* - Updated submission tests
- Moved create_datafile to util

* - fix lint errors

* - removing frontend filtering

* - addding datafile to admin model

* Revert "- addding datafile to admin model"

This reverts commit 35a6f24c36c3a4c00ddcfc40f20833530b0199f4.

* - Fixed issue where datafile FK wasnt populating
- Regenerated migration

* - Readding datafile back to admin view now that the error is resolved

* - adding datafile back

* Revert "- Readding datafile back to admin view now that the error is resolved"

This reverts commit 2807425059fd1b5b355edfb16d30d170cf869d7b.

* - Removed unnecessary fields
- Updated dependencies
- Updated filter

* - Updated document to include required fields

* - Moved datafile reference to avoid confusion

* - Updating based on comments.
- Added base class to keep things a little more DRY

* - Refactored parsing to avoid custom transactions by leveraging the FK on the records. Rollbacks are a lot slower now, but hopefully that will happen much less versus successful parses.

* - small optimization for the rollback

* - Fix lint errors

* - Removing commented code

* - Installing build essentials

* Revert "- Installing build essentials"

This reverts commit 782a5f7d687e60965a9100520f027b9678dbac35.

* - adding setup tools and wheel

* Revert "- adding setup tools and wheel"

This reverts commit f529728811fba242132b7c42f9e9e09d6037fa70.

* - Updating dependencies to get around pep issue

* - Pin factoryboy
- fix lint error

* - Updating manifest

* - Added EncryptedField class
- Updated schema's
- Updated datafiles since all are encrypted

* - Fix lint errors

* - Added decryption for ssp

* - Making the encrypted check stronger

* - Refactored fields and rowschema into seperate files
- Updated encrypted field to take decryption function

* - fix lint errors

* - Fix import error

* - Removed erroneous update that was causing error
- Fixed whitespace in datafiles

* - White space error

* - Resolved error

* - Refactored encryption logic to take the header encryption indicator into account

* - Fix lint errors

---------
