+++ tdrs-backend/tdpservice/parsers/schema_defs/ssp/m1.py
@@ -1,7 +1,9 @@
 """Schema for SSP M1 record type."""
 
 
-from ...util import RowSchema, Field, SchemaManager
+from ...util import SchemaManager
+from ...fields import Field
+from ...row_schema import RowSchema
 from ... import validators
 from tdpservice.search_indexes.models.ssp import SSP_M1
