+++ tdrs-backend/tdpservice/parsers/schema_defs/tanf/t2.py
@@ -1,7 +1,9 @@
 """Schema for HEADER row of all submission types."""
 
 
-from ...util import RowSchema, Field, SchemaManager
+from ...util import SchemaManager
+from ...fields import EncryptedField, Field, tanf_ssn_decryption_func
+from ...row_schema import RowSchema
 from ... import validators
 from tdpservice.search_indexes.models.tanf import TANF_T2
 
@@ -27,8 +29,8 @@ t2 = SchemaManager(
                     required=True, validators=[]),
               Field(item=6, name='DATE_OF_BIRTH', type='number', startIndex=21, endIndex=29,
                     required=True, validators=[]),
-              Field(item=7, name='SSN', type='string', startIndex=29, endIndex=38,
-                    required=True, validators=[]),
+              EncryptedField(decryption_func=tanf_ssn_decryption_func, item=7, name='SSN', type='string', startIndex=29,
+                             endIndex=38, required=True, validators=[]),
               Field(item=8, name='RACE_HISPANIC', type='string', startIndex=38, endIndex=39,
                     required=True, validators=[]),
               Field(item=9, name='RACE_AMER_INDIAN', type='string', startIndex=39, endIndex=40,
