+++ tdrs-backend/tdpservice/parsers/schema_defs/trailer.py
@@ -1,7 +1,8 @@
 """Schema for TRAILER row of all submission types."""
 
 
-from ..util import RowSchema, Field
+from ..fields import Field
+from ..row_schema import RowSchema
 from .. import validators
