+++ tdrs-backend/tdpservice/parsers/test/data/ADS.E2J.FTP1.TS06
@@ -1,4 +1,4 @@
-HEADER20204A06   TAN1 N
+HEADER20204A06   TAN1EN
 T12020101111111111223003403361110213120000300000000000008730010000000000000000000000000000000000222222000000002229012                                       
 T2202010111111111121219740114WTTTTTY@W2221222222221012212110014722011400000000000000000000000000000000000000000000000000000000000000000000000000000000000291
 T320201011111111112120190127WTTTT90W022212222204398100000000
